+++ src/Data/SkillbarData.h
@@ -1,5 +1,7 @@
 #pragma once
 
+#include "stdafx.h"
+
 #include <array>
 #include <cstdint>
 #include <string_view>
