+++ src/Helper/HelperUw.cpp
@@ -8,16 +8,12 @@
 
 #include <Helper.h>
 #include <HelperPlayer.h>
+#include <HelperUwPos.h>
 #include <MathUtils.h>
 #include <PlayerData.h>
 #include <Types.h>
 
 #include "HelperUw.h"
-
-bool IsUw()
-{
-    return GW::Map::GetMapID() == GW::Constants::MapID::The_Underworld;
-}
 
 bool UwHelperActivationConditions()
 {
@@ -28,6 +24,89 @@ bool UwHelperActivationConditions()
         return false;
 
     return true;
+}
+
+
+uint32_t GetTankId()
+{
+    std::vector<PlayerMapping> party_members;
+    const auto success = GetPartyMembers(party_members);
+
+    if (!success || party_members.size() < 2)
+        return 0;
+
+    auto tank_idx = uint32_t{0};
+
+    switch (GW::Map::GetMapID())
+    {
+    case GW::Constants::MapID::The_Underworld:
+#ifdef _DEBUG
+    case GW::Constants::MapID::Isle_of_the_Nameless:
+#endif
+    {
+        tank_idx = party_members.size() - 2;
+        break;
+    }
+    default:
+    {
+        tank_idx = 0;
+        break;
+    }
+    }
+
+    const auto tank = party_members[tank_idx];
+    return tank.id;
+}
+
+uint32_t GetEmoId()
+{
+    std::vector<PlayerMapping> party_members;
+    const auto success = GetPartyMembers(party_members);
+
+    if (!success)
+        return 0;
+
+    for (const auto &member : party_members)
+    {
+        const auto agent = GW::Agents::GetAgentByID(member.id);
+        if (!agent)
+            continue;
+
+        const auto living = agent->GetAsAgentLiving();
+        if (!living)
+            continue;
+
+        if (living->primary == static_cast<uint8_t>(GW::Constants::Profession::Elementalist) &&
+            living->secondary == static_cast<uint8_t>(GW::Constants::Profession::Monk))
+            return agent->agent_id;
+    }
+
+    return 0;
+}
+
+uint32_t GetDhuumBitchId()
+{
+    std::vector<PlayerMapping> party_members;
+    const auto success = GetPartyMembers(party_members);
+
+    if (!success)
+        return 0;
+
+    for (const auto &member : party_members)
+    {
+        const auto agent = GW::Agents::GetAgentByID(member.id);
+        if (!agent)
+            continue;
+
+        const auto living = agent->GetAsAgentLiving();
+        if (!living)
+            continue;
+
+        if (living->secondary == static_cast<uint8_t>(GW::Constants::Profession::Ranger))
+            return agent->agent_id;
+    }
+
+    return 0;
 }
 
 bool IsEmo(const PlayerData &player_data)
@@ -83,98 +162,6 @@ bool IsMesmerTerra(const PlayerData &player_data)
         return false;
 
     return !IsLT(player_data);
-}
-
-bool IsAtSpawn(const GW::GamePos &player_pos)
-{
-    return IsNearToGamePos(player_pos, GW::GamePos{1248.00F, 6965.51F, 0}, 500.0F);
-}
-
-bool IsAtChamberSkele(const GW::GamePos &player_pos)
-{
-    return IsNearToGamePos(player_pos, GW::GamePos{-2726.856F, 10239.48F, 0}, 2250.0F);
-}
-
-bool IsAtBasementSkele(const GW::GamePos &player_pos)
-{
-    return IsNearToGamePos(player_pos, GW::GamePos{-5183.64F, 8876.31F, 0}, 2000.0F);
-}
-
-bool IsRightAtChamberSkele(const GW::GamePos &player_pos)
-{
-    return IsNearToGamePos(player_pos, GW::GamePos{-2726.856F, 10239.48F, 0}, 300.0F);
-}
-
-bool IsAtFusePull1(const GW::GamePos &player_pos)
-{
-    return IsNearToGamePos(player_pos, GW::GamePos{-6263.33F, 9899.79F, 0}, 1500.0F);
-}
-
-bool IsAtFusePull2(const GW::GamePos &player_pos)
-{
-    return IsNearToGamePos(player_pos, GW::GamePos{-7829.98F, 4324.09F, 0}, 1500.0F);
-}
-
-bool IsAtFusePulls(const GW::GamePos &player_pos)
-{
-    if (!IsUw())
-        return false;
-
-    if (IsAtFusePull1(player_pos) || IsAtFusePull2(player_pos))
-        return true;
-
-    return false;
-}
-
-bool IsAtValeStart(const GW::GamePos &player_pos)
-{
-    return IsNearToGamePos(player_pos, GW::GamePos{-9764.08F, 2056.60F, 0}, 1500.0F);
-}
-
-bool IsAtValeHouse(const GW::GamePos &player_pos)
-{
-    return IsNearToGamePos(player_pos, GW::GamePos{-12264.12F, 1821.18F, 0}, 1500.0F);
-}
-
-bool IsRightAtValeHouse(const GW::GamePos &player_pos)
-{
-    return IsNearToGamePos(player_pos, GW::GamePos{-12264.12F, 1821.18F, 0}, 5.0F);
-}
-
-bool IsAtSpirits1(const GW::GamePos &player_pos)
-{
-    return IsNearToGamePos(player_pos, GW::GamePos{-13872.34F, 2332.34F, 0}, GW::Constants::Range::Spellcast);
-}
-
-bool IsAtSpirits2(const GW::GamePos &player_pos)
-{
-    return IsNearToGamePos(player_pos, GW::GamePos{-13760.19F, 358.15F, 0}, GW::Constants::Range::Spellcast);
-}
-
-bool IsAtValeSpirits(const GW::GamePos &player_pos)
-{
-    if (!IsUw())
-        return false;
-
-    if (IsAtSpirits1(player_pos) || IsAtSpirits2(player_pos))
-        return true;
-
-    return false;
-}
-
-bool IsInDhuumRoom(const GW::GamePos &player_pos)
-{
-#ifdef _DEBUG
-    if (GW::Map::GetMapID() == GW::Constants::MapID::Isle_of_the_Nameless)
-        return true;
-#endif
-
-    return IsNearToGamePos(player_pos, GW::GamePos{-16105.50F, 17284.84F, 0}, GW::Constants::Range::Spellcast);
-}
-
-bool IsGoingToDhuum(const GW::GamePos &player_pos)
-{
-    return IsNearToGamePos(player_pos, GW::GamePos{-9567.56F, 17288.916F, 0}, 100.0F);
 }
 
 bool IsInDhuumFight(uint32_t *dhuum_id, float *dhuum_hp, uint32_t *dhuum_max_hp)
@@ -280,7 +267,7 @@ bool TargetIsReaper(PlayerData &player_data)
 
 bool TargetReaper(PlayerData &player_data)
 {
-    return TargetClosestNpcById(player_data, GW::Constants::ModelID::UW::Reapers) != 0;
+    return TargetClosestNpcById(player_data, GW::Constants::ModelID::UW::Reapers) != 0U;
 }
 
 bool TalkReaper(PlayerData &player_data)
@@ -294,7 +281,7 @@ bool TalkReaper(PlayerData &player_data)
     if (!agent)
         return true;
 
-    GW::Agents::GoNPC(agent, 0);
+    GW::Agents::GoNPC(agent, 0U);
 
     return true;
 }
@@ -376,86 +363,4 @@ bool DhuumIsCastingJudgement(const uint32_t dhuum_id)
         return true;
 
     return false;
-}
-
-uint32_t GetTankId()
-{
-    std::vector<PlayerMapping> party_members;
-    const auto success = GetPartyMembers(party_members);
-
-    if (!success || party_members.size() < 2)
-        return 0;
-
-    auto tank_idx = uint32_t{0};
-
-    switch (GW::Map::GetMapID())
-    {
-    case GW::Constants::MapID::The_Underworld:
-#ifdef _DEBUG
-    case GW::Constants::MapID::Isle_of_the_Nameless:
-#endif
-    {
-        tank_idx = party_members.size() - 2;
-        break;
-    }
-    default:
-    {
-        tank_idx = 0;
-        break;
-    }
-    }
-
-    const auto tank = party_members[tank_idx];
-    return tank.id;
-}
-
-uint32_t GetEmoId()
-{
-    std::vector<PlayerMapping> party_members;
-    const auto success = GetPartyMembers(party_members);
-
-    if (!success)
-        return 0;
-
-    for (const auto &member : party_members)
-    {
-        const auto agent = GW::Agents::GetAgentByID(member.id);
-        if (!agent)
-            continue;
-
-        const auto living = agent->GetAsAgentLiving();
-        if (!living)
-            continue;
-
-        if (living->primary == static_cast<uint8_t>(GW::Constants::Profession::Elementalist) &&
-            living->secondary == static_cast<uint8_t>(GW::Constants::Profession::Monk))
-            return agent->agent_id;
-    }
-
-    return 0;
-}
-
-uint32_t GetDhuumBitchId()
-{
-    std::vector<PlayerMapping> party_members;
-    const auto success = GetPartyMembers(party_members);
-
-    if (!success)
-        return 0;
-
-    for (const auto &member : party_members)
-    {
-        const auto agent = GW::Agents::GetAgentByID(member.id);
-        if (!agent)
-            continue;
-
-        const auto living = agent->GetAsAgentLiving();
-        if (!living)
-            continue;
-
-        if (living->secondary == static_cast<uint8_t>(GW::Constants::Profession::Ranger))
-            return agent->agent_id;
-    }
-
-    return 0;
 }
