+++ src/Main/Features/AutoFollowWindow.cpp
@@ -1,5 +1,3 @@
-#include "stdafx.h"
-
 #include <array>
 #include <cstdint>
